requirements.txt
fastapi
uvicorn
pandas
scikit-learn==1.8.0
mlflow
sentence-transformers
torch
prometheus-client
joblib
kafka-python
psycopg2-binary
python-multipart
